fix: invalid query on mysql

+++ backend/src/main/java/com/maplewood/infrastructure/persistence/repository/StudentAcademicMetricsProjection.java
@@ -1,7 +1,9 @@
 package com.maplewood.infrastructure.persistence.repository;
 
+import java.math.BigDecimal;
+
 public interface StudentAcademicMetricsProjection {
-    Integer getCreditsEarned();
+    BigDecimal getCreditsEarned();
 
     Double getCalculatedGpa();
 }
